research/requirements.txt
setuptools==78.1.0
Flask==3.1.0
prometheus-flask-exporter==0.23.2
python-dotenv==1.1.0
flask-cors==5.0.1
gunicorn==23.0.0
uuid-v7==1.0.0
# ORM dependencies
azure-storage-blob==12.25.1
redis==5.2.1
kafka-python==2.1.5
psycopg2-yugabytedb==2.9.3.5
Flask-SQLAlchemy==3.1.1
# For paper scraping
paperscraper==0.3.0
nltk==3.9.1
pymupdf==1.25.5
pymupdf4llm==0.0.20
# Added RAG dependencies
openai==1.70.0
langchain==0.3.23
langchain-core==0.3.51
langchain-community==0.3.21
langchain-text-splitters==0.3.8
langchain-deepseek==0.1.3
langchain-openai==0.3.12
langchain-ollama==0.3.0
langchain-postgres==0.0.13
# Websearch dependencies
duckduckgo-search==7.5.3
crawl4ai==0.5.0post8